apps/marketdata/requirements.txt
appdirs==1.4.3
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.8
pathspec==0.7.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
regex==2020.1.8
requests==2.22.0
rsa==4.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
